refac: [코드리뷰 반영]  ApiPath 제거 및 RestAssuredClient 엔티티별로 분리
