--- src/test/java/subway/common/ApiPath.java
@@ -1,15 +1,0 @@
-package subway.common;
-
-public class ApiPath {
-    // Station
-    public static final String STATION_CREATE_PATH = "/stations";
-    public static final String STATION_DELETE_PATH = "/stations/%d";
-    public static final String STATION_LIST_PATH = "/stations";
-
-    // Line
-    public static final String LINE_CREATE_PATH = "/lines";
-    public static final String LINE_GET_PATH = "/lines/%d";
-    public static final String LINE_UPDATE_PATH = "/lines/%d";
-    public static final String LINE_LIST_PATH = "/lines";
-    public static final String LINE_DELETE_PATH = "/lines/%d";
-}
